requirements.txt
openai>=1.52.0
rich>=13.7.1
click>=8.1.7
python-dotenv>=1.0.1
